Don't raise error when configured before database schema loaded
To setup our store, we need to inspect the original store column and
then "wrap" it in our own deserialization logic (which unpacks the
attributes following their own deserialization rules).

However, we can lazily evaluate this wrapper. Rails does this using a
`attributes_to_define_after_schema_loads` array, which is invoked when
the schema is loaded. Using the block form means that the type of the
store column is only loaded when the schema is loaded, and in the
meantime the code can be used.

This fixes, for example, checking the store attributes during asset
compilation when a database is not available.