Prevent enums being combined with array: true
Currently this feature cannot work because of how enums are implemented
using validations. The validator assumes a single field, for array mode
it would need to understand how to check each element (this is something
that ActiveModel::Validations does not yet support).

Worse, however, is that arrays are implemented by _wrapping_ the
original type, so you would have ArrayWrapper.new(Types::Enum.new(...)).
This means that the inner Enum type does not know that it is wrapped in
an array, so it could not modify its behaviour properly.

This feature might be added in future, but it would require some
reworking of the internal implementation. In the meantime, this
exception protects users from believing their enum is working correctly.