Refactor internals with AttributeSet to fix typecasting woes
This commit is a sizeable refactor of this gem to reuse some of the core
features from ActiveModel::AttributeSet. The key change is to track
changes as Attribute objects, rather than their underlying values. This
allows us (and Rails) to determine whether a change has come from the
user, or the database, and thus what typecasting to apply at that point.

The complexity in the gem has shifted now from the getter/setter methods
for each attribute, onto the StoreColumnWrapper class which translates
the actual column in Rails into a new AttributeSet object. As properties
are mutated, the AttributeSet tracks the changes, and when the object is
serialized for the database it takes care of all the ugly serialization
parts.

Unfortunately, we had to lose two (as far as I can tell unused)
features, blocks with default values and arrays with default values that
wasn't an empty array.

We had to lose the block support because the default value is now
calculated by the StoreColumnWrapper#deserialize method, and at that
point in the code it does not have access to the underlying record
object (this is a Rails design presumably to enforce modularity).
Default values cannot be calculated without that context.

Array defaults meanwhile turned out to be rather difficult to retain
when AttributeSet is handling change tracking for us. The sentinel
"EMPTY" value required a lot of modification of pretty much every method
to accommodate; it didn't seem worth the carrying cost.

However, despite these two feature losses, the big major win is that you
can now read back a value immediately after writing it (before
persisting it to the DB). This replaces all of the "bad_typecasting"
logic that we had before.