Fix dirty tracking after loading clean record
To trigger this bug, we need to load the properties set and try and read
one of the values back. This triggers the query on
`ActiveModel::Attribute#has_been_read?`, and the logic for checking
whether the object has been mutated becomes more involved.

To fix it, we need to:

  * Actually deserialize raw_original_value (it's a JSON string usually)
  * Check both attribute sets more carefully
  * Check if any of the types have been changed "in place", which is
    true if you modify the record without a usual assignment, especially
    on arrays, e.g:

       record.listy << "foo"