Include possible enum values in enum validation error message
Rails does not reveal what the possible values are for an enum by
default. This isn't explained in the source code, so I can see two
possible reasons for this:

  1. Security through obscurity: don't reveal possible values and it
     reduces an attack vector where somebody abuses a column with a
     magic value (e.g. `role = superadmin`)

  2. The `in:` option supports `Range`, which means you could
     potentially have something like `in: (1..100_000)`. Printing that
     as a list of possible options would be a bad choice...

However, the enums we're working with in this gem are usually a much
smaller set, and could really provide a better user experience. So this
commit adds a validator which extends the normal inclusion validator,
but also injects the possible values in the message.

Turns out the hardest part was injecting the translations. With a
Rails::Engine-style gem, this gets handled automatically, but here I had
to use a Railtie just to get Rails to recognise them (this is important,
if you just do `I18n.load_path << ...` when Rails loads itself it
_overwrites_ whatever translation files you set 🙈